bic-analysis-main/requirements.txt
pandas==2.2.3
streamlit==1.44.0
numpy==1.26.2
openai==1.63.2
matplotlib==3.9.2
seaborn==0.13.2
scikit-learn==1.5.2
SQLAlchemy==2.0.23
python-dotenv==1.0.1
PyMySQL==1.1.1
statsmodels==0.14.4
bcrypt==4.3.0